add: partial serde support
This change adds support for most of the serde data types except floats.

There are tests in place for when floats are added but they have extra
parsing complications so they will be added in a separate patch.

This PR moves the original tnetstring parsing to a new file and ends up
not using the parsing for the serde functionality.